update with delete many and validation

+++ src/homeContents/homeContents.service.ts
@@ -17,8 +17,11 @@ export class HomeContentsService {
     createHomeContentsDTO: CreateHomeContentsDTO,
     image: Express.Multer.File,
   ) {
-    const imagePathProccess = await this.imageUploadService.imageUpload(image);
-    createHomeContentsDTO.image = imagePathProccess;
+    if (image !== undefined) {
+      const imagePathProccess =
+        await this.imageUploadService.imageUpload(image);
+      createHomeContentsDTO.image = imagePathProccess;
+    }
     const newHomeContent = new this.homeContentModel(createHomeContentsDTO);
     return await newHomeContent.save();
   }
@@ -48,6 +51,11 @@ export class HomeContentsService {
     updateHomeContentsDTO: UpdateHomeContentsDTO,
     image: Express.Multer.File,
   ) {
+    const findData = await this.homeContentModel.findById(id);
+    const imagePath = findData?.image.toString();
+    if (typeof imagePath === 'string' && imagePath !== '') {
+      fs.unlink(imagePath);
+    }
     const imagePathProcess = await this.imageUploadService.imageUpload(image);
     updateHomeContentsDTO.image = imagePathProcess;
     return await this.homeContentModel.findByIdAndUpdate(
@@ -61,7 +69,7 @@ export class HomeContentsService {
     try {
       const findData = await this.homeContentModel.findById(id);
       const imagePath = findData?.image.toString();
-      if (imagePath) {
+      if (typeof imagePath === 'string' && imagePath !== '') {
         fs.unlink(imagePath);
         console.log('Image delete');
       }
@@ -72,4 +80,25 @@ export class HomeContentsService {
   }
 
   //Delete many home content
+  async deleteManyHomeContent(ids: string[]) {
+    try {
+      const isValidPath = (path: string) => {
+        return typeof path === 'string' && path.trim() !== '';
+      };
+
+      const manyData = await this.homeContentModel.find({
+        _id: { $in: ids },
+      });
+      for (let i = 0; i < manyData.length; i++) {
+        if (isValidPath(manyData[i].image)) {
+          fs.unlink(manyData[i].image);
+        }
+      }
+      return await this.homeContentModel.deleteMany({
+        _id: { $in: ids },
+      });
+    } catch (error) {
+      console.error(error);
+    }
+  }
 }
